requirements.txt
adjustText==0.7.3
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
biothings-client==0.3.0
blessed==1.20.0
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
cloudpickle==2.2.1
comm==0.1.3
contourpy==1.0.7
cubinlinker-cu11==0.3.0.post1
cuda-python==11.8.2
cudf-cu11==23.10.0
cuml-cu11==23.10.0
cupy-cuda11x==12.2.0
cycler==0.11.0
Cython==0.29.35
dask==2023.9.2
dask-cuda==23.10.0
dask-cudf-cu11==23.10.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
distributed==2023.9.2
exceptiongroup==1.1.1
executing==1.2.0
fastrlock==0.8.2
filelock==3.9.0
fonttools==4.39.4
fsspec==2023.5.0
gpustat==1.1.1
graphviz==0.20.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.23.2
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.1
kiwisolver==1.4.4
llvmlite==0.40.1
locket==1.0.0
MarkupSafe==2.1.3
matplotlib==3.3.4
matplotlib-inline==0.1.6
matplotlib-venn==0.11.9
mpmath==1.3.0
msgpack==1.0.7
mygene==3.2.2
nest-asyncio==1.5.6
networkx==3.1
numba==0.57.1
numpy==1.24.4
nvidia-ml-py==12.535.108
nvtx==0.2.8
packaging==23.1
pandas==1.5.3
parso==0.8.3
partd==1.4.0
patsy==0.5.3
pexpect==4.8.0
Pgenlib==0.83.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==4.24.4
psutil==5.9.5
ptxcompiler-cu11==0.7.0.post1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
Pygments==2.15.1
pylibraft-cu11==23.10.0
pynvml==11.4.1
pyparsing==3.0.9
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
pyzmq==25.1.0
raft-dask-cu11==23.10.0
requests==2.31.0
rmm-cu11==23.10.0
scikit-learn==0.24.2
scipy==1.11.3
seaborn==0.12.2
shap==0.42.1
six==1.16.0
slicer==0.0.7
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
statsmodels==0.12.2
sympy==1.12
tblib==2.0.0
threadpoolctl==3.1.0
tomli==2.0.1
toolz==0.12.0
torch==2.1.0+cu118
torchviz==0.0.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
treelite==3.9.1
treelite-runtime==3.9.1
triton==2.1.0
typing_extensions==4.6.3
tzdata==2023.3
ucx-py-cu11==0.34.0
UpSetPlot==0.8.0
urllib3==2.0.4
venn==0.1.3
wcwidth==0.2.6
wrapt==1.15.0
zict==3.0.0
zipp==3.15.0
